final Integer quota=(Integer)is.getObject("quota");

+++ opensso/products/amserver/source/org/forgerock/openam/session/service/DestroyAllAction.java
@@ -52,7 +52,7 @@ public class DestroyAllAction implements QuotaExhaustionAction {
 
     @Override
     public boolean action(InternalSession is, Map sessions) {
-    	final Long quota=(Long)is.getObject("quota");
+    	final Integer quota=(Integer)is.getObject("quota");
     	
     	if (sessions.size()<quota)
     		return false;
