+++ opensso/products/amserver/source/org/forgerock/openam/session/service/DestroyOldestAction.java
@@ -67,7 +67,7 @@ public class DestroyOldestAction implements QuotaExhaustionAction {
 	
     @Override
     public boolean action(InternalSession is, Map<String, Long> sessions) {
-    	final Long quota=(Long)is.getObject("quota");
+    	final Integer quota=(Integer)is.getObject("quota");
     	
     	if (sessions.size()<quota)
     		return false;
